[common/meta/embedded] refactor: move kv_api impl into standalone file
